+++ frontend/.next/server/webpack-runtime.js
@@ -125,7 +125,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("77fa5e5d0df18df5")
+/******/ 		__webpack_require__.h = () => ("ba22f51685e79530")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/hasOwnProperty shorthand */