+++ frontend/.next/static/chunks/webpack.js
@@ -192,7 +192,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	!function() {
-/******/ 		__webpack_require__.h = function() { return "cca9b0e65ab0876d"; }
+/******/ 		__webpack_require__.h = function() { return "39f796a972f02d7a"; }
 /******/ 	}();
 /******/ 	
 /******/ 	/* webpack/runtime/global */